renamed sliderbar to navbar and other changes removed
