+++ SpringBootServer/IndiSky/src/main/java/com/indisky/user/controller/FlightController.java
@@ -14,7 +14,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @RestController
-//@CrossOrigin(origins = "http://localhost:5173")
+@CrossOrigin(origins = "http://localhost:5173")
 @AllArgsConstructor
 @RequestMapping("/api/flights")
 public class FlightController {
