+++ SpringBootServer/IndiSky/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/gauravtest?createDatabaseIfNotExist=true
+spring.datasource.url=jdbc:mysql://localhost:3306/indisky?createDatabaseIfNotExist=true
 spring.jpa.hibernate.ddl-auto=update
 
 #Akash :
@@ -6,8 +6,8 @@ spring.jpa.hibernate.ddl-auto=update
 #spring.datasource.password=manager
 
 #Piyush :
-#spring.datasource.username=root
-#spring.datasource.password=Pass@123
+spring.datasource.username=root
+spring.datasource.password=Pass@123
 
 #Kalpesh :
 #spring.datasource.username=root
