+++ client/src/Pages/BookingConfirmation.jsx
@@ -3,7 +3,7 @@ import SlideBar from '../Component/SlideBar';
 import Footer from '../Component/Footer';
 import { useNavigate } from 'react-router-dom';
 import '../css/StaticPage.css'; // ✅ Reuse shared styles
-import Sidebar from '../Component/Sidebar';
+import Sidebar from '../Component/Navbar';
 
 function BookingConfirmation() {
   const navigate = useNavigate();
